feat(eslint-plugin-snacks-pie-migration): DSW-3604 export component data extractor
